Add timeout around request / response id matching while loop:
- Use the call timeout for the provider to set up a reasonable escape time for the while look that tries to match the response `id` with the request `id`. All providers should technically return matching ids in the response but we've run into issues with this in the past and this prevents an infinite loop, handling it somewhat gracefully.
- Make sure that the ``WebsocketProviderV2`` sets a default timeout.

+++ web3/providers/persistent.py
@@ -34,7 +34,7 @@ class PersistentConnectionProvider(AsyncJSONBaseProvider, ABC):
         self,
         endpoint_uri: str,
         request_cache_size: int = 100,
-        call_timeout: int = DEFAULT_PERSISTENT_CONNECTION_TIMEOUT,
+        call_timeout: float = DEFAULT_PERSISTENT_CONNECTION_TIMEOUT,
     ) -> None:
         super().__init__()
         self.endpoint_uri = endpoint_uri
